+++ Jarvis.Framework.Kernel/ProjectionEngine/Rebuild/RebuildProjectionMetrics.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Jarvis.Framework.Kernel.ProjectionEngine.Rebuild
+{
+    internal static class RebuildProjectionMetrics
+    {
+#pragma warning disable S2223 // Non-constant static fields should not be visible
+        internal static Int32 CountOfConcurrentDispatchingCommit = 0;
+#pragma warning restore S2223 // Non-constant static fields should not be visible
+    }
+}